[update]
-refactoring kafka admin/producer/consumer client.
-update kafka admin/producer/consumer actor client.

+++ src/test/scala/com/example/kafka/producer/TestProducerClient.scala
@@ -2,7 +2,6 @@ package com.example.kafka.producer
 
 import com.example.kafka.admin.AdminClient
 import org.apache.kafka.clients.producer.{ProducerRecord, RecordMetadata}
-import org.apache.kafka.common.serialization.StringSerializer
 import org.junit._
 import org.hamcrest.CoreMatchers._
 import com.example.kafka.producer.TestProducerClient._
@@ -11,17 +10,9 @@ import com.example.utils.AppConfig
 class TestProducerClient {
 
   @Test
-  def testProduceRecord(): Unit = {
-    val result: RecordMetadata = testProducerClient.produceRecord(testProduceRecordSet(0)).get()
-    val resultMetaData: Map[String, Any] = ProducerClient.produceRecordMetadataToMap(result)
-
-    Assert.assertThat(resultMetaData("topic").toString, is(testTopicName))
-  }
-
-  @Test
   def testProduceRecords(): Unit = {
     val result: Vector[RecordMetadata] = testProducerClient.produceRecords(testProduceRecordSet).map(_.get)
-    val resultMetaData: Vector[Map[String, Any]] = result.map(ProducerClient.produceRecordMetadataToMap)
+    val resultMetaData: Vector[Map[String, String]] = result.map(ProducerClient.produceRecordMetadataToMap)
 
     Assert.assertThat(resultMetaData.length, is(testProduceRecordSetCount))
     Assert.assertThat(resultMetaData.count(_("topic").toString == testTopicName), is(testProduceRecordSetCount))
@@ -39,30 +30,39 @@ object TestProducerClient {
       new ProducerRecord(testTopicName, s"key-$i".asInstanceOf[Any], s"value-$i".asInstanceOf[Any])
     }.toVector
 
-  val testAdminClient = AdminClient(AppConfig.getKafkaAdminProps)
-
-  var testProducerClient: ProducerClient[Any, Any] = _
-
+  val testAdminClient = AdminClient(AppConfig.createDefaultKafkaAdminProps)
+  var testProducerClient: ProducerClient[Any, Any] = ProducerClient[Any, Any](AppConfig.createDefaultKafkaProducerProps)
 
   @BeforeClass
   def beforeClass(): Unit = {
-    testProducerClient = ProducerClient(AppConfig.getKafkaProducerProps)
-
-    while(!testAdminClient.isExistTopic(testTopicName)) {
-      testAdminClient.createTopic(testTopicName, testTopicPartitionCount, testTopicReplicationFactor).get
-      Thread.sleep(500)
-    }
+    this.createTestTopic()
   }
 
   @AfterClass
   def tearDownClass(): Unit = {
+    this.deleteTestTopic()
+
     testProducerClient.close()
-
-    while(testAdminClient.isExistTopic(testTopicName)) {
-      testAdminClient.deleteTopic(testTopicName).get
-      Thread.sleep(500)
-    }
-
     testAdminClient.close()
+  }
+
+  def createTestTopic(): Unit = {
+    if (!testAdminClient.isExistTopic(testTopicName)) {
+      testAdminClient.createTopic(testTopicName, testTopicPartitionCount, testTopicReplicationFactor).get
+      while(!testAdminClient.isExistTopic(testTopicName)) {
+        testAdminClient.createTopic(testTopicName, testTopicPartitionCount, testTopicReplicationFactor).get
+        Thread.sleep(500)
+      }
+    }
+  }
+
+  def deleteTestTopic(): Unit = {
+    if (testAdminClient.isExistTopic(testTopicName)) {
+      testAdminClient.deleteTopic(testTopicName).get
+      while(testAdminClient.isExistTopic(testTopicName)) {
+        testAdminClient.deleteTopic(testTopicName).get
+        Thread.sleep(500)
+      }
+    }
   }
 }
